backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM
anthropic==0.18.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
ruff==0.1.14

# Utilities
python-dotenv==1.0.0
